Adapt the code to the required changes
