+++ functions/src/modules/email.functions.ts
@@ -14,6 +14,81 @@ let emailTemplates = {};
 
 
 
+
+// SendInBlue Helper Functions
+import * as SibApiV3Sdk from 'sib-api-v3-sdk';
+const sendInBlueSecretKey = functions.config().send_in_blue.key;
+const defaultClient = SibApiV3Sdk.ApiClient.instance;
+let apiKey = defaultClient.authentications['api-key'];
+apiKey.apiKey = sendInBlueSecretKey;
+
+const apiInstance = new SibApiV3Sdk.SMTPApi();
+
+export const sendEmail = async (to, bcc, replyTo, templateId, params) => {
+    let smtpEmail = new SibApiV3Sdk.SendSmtpEmail();
+
+    smtpEmail = {
+        to: [{ email: to }],
+        bcc,
+        replyTo,
+        templateId,
+        params
+    };
+
+    let sendingResult = await apiInstance.sendTransacEmail(smtpEmail);
+    console.log(sendingResult);
+}
+
+export const updateTemplate = async (templateId, html) => {
+    let updatedTemplate = new SibApiV3Sdk.CreateSmtpTemplate();
+
+    updatedTemplate.htmlContent = html;
+
+    let updateResult = await apiInstance.updateSmtpTemplate(templateId, updatedTemplate);
+    console.log(updateResult);
+
+}
+
+
+
+export const createTemplate = async (templateName, sender, html, subject) => {
+    let smtpTemplate = new SibApiV3Sdk.CreateSmtpTemplate();
+
+    smtpTemplate.templateName = templateName;
+    smtpTemplate.htmlContent = html;
+    smtpTemplate.subject = subject;
+    smtpTemplate.sender = sender;
+    smtpTemplate.isActive = true;
+
+    let result = await apiInstance.createSmtpTemplate(smtpTemplate);
+    return result['id'];
+}
+
+
+export const getTemplateId = async (templateName) => {
+    var opts = { 'templateStatus': true };
+    
+    let result = await apiInstance.getSmtpTemplates(opts)
+    let { templates } = result;
+
+    let template = templates.filter(temp => temp['name'] === templateName)[0]
+
+    if (template === undefined)
+        return -1; // template not found
+    else
+        return templates['id'];
+}
+
+
+export const sendTestTemplate = async (templateId) => {
+    let testEmail = new SibApiV3Sdk.SendTestEmail();
+    await apiInstance.sendTestTemplate(templateId, testEmail);
+
+}
+
+
+
+
 /////////////////////////////////////////////////
 //          Update Email Templates on sendInBlue 
 //          in response to 
@@ -52,21 +127,21 @@ const updateEmailTemplates = async (filePath) => {
     
     let { sender, subject } = template;
 
-    let id = helpers.getTemplateId(fileName);
+    let id = getTemplateId(fileName);
     
     if (+id === -1) { // New Template
-        id = await helpers.createTemplate(fileName, sender, htmlContent, subject);
+        id = await createTemplate(fileName, sender, htmlContent, subject);
         await db.ref(`/email/templates/${fileName}`)
                 .update({ lastUpdated: new Date() });
     }            
     else {
-        await helpers.updateTemplate(id, htmlContent);
+        await updateTemplate(id, htmlContent);
         await db.ref(`/email/templates/${template.key}`)
                 .update({ lastUpdated: new Date() });
     }
 
     // 2. Send a test Email confirming it has been updated correctly
-    helpers.sendTestTemplate(id, sender.email);        
+    sendTestTemplate(id);        
 }
 
 export const updateTemplatesOnTemplateUpload = functions.storage.object()
@@ -92,9 +167,10 @@ export const sendNotificationEmail = functions.database.ref('/email/notification
     if (Object.keys(emailTemplates).indexOf(templateName) > -1)
         id = emailTemplates[templateName].id;
     else 
-        id = await helpers.getTemplateId(templateName)
+        id = await getTemplateId(templateName)
 
-    await helpers.sendEmail(data.to, data.bcc, id, data.params);
+    if (!data['sentTimestamp'])
+        await sendEmail(data.to, data.bcc, data.replyTo, id, data.params);
 
-    return snapshot.ref.update({ sent: true });
+    return snapshot.ref.update({ sentTimestamp: Math.round((new Date()).getTime() / 1000) });
 });
